% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crochet.R
\name{crochet}
\alias{crochet}
\title{Creates an implementation of [ for custom matrix-like types}
\usage{
crochet(subset_vector, subset_matrix)
}
\arguments{
\item{subset_vector}{A function in the form of \code{function(x, i)} that takes a
subset of \code{x} based on a single index \code{i} and returns a vector.}

\item{subset_matrix}{A function in the form of \code{function(x, i, j)} that
takes a subset of \code{x} based on two indices \code{i} and \code{j} and returns a matrix.}
}
\value{
A function in the form of \code{function(x, i, j, drop = TRUE} that is
meant to be used as a method for \code{\link[base]{[}}.
}
\description{
crochet is a function that accepts two arguments \code{subset_vector} (in the
form of \code{function(x, i)}) and \code{subset_matrix} (in the form of \code{function(x, i, j)}), and returns a function that can be used as a method for
\code{\link[base]{[}} for a custom type.
}
\details{
The custom type must implement methods for \code{\link[base:dim]{base::dim()}} and
\code{\link[base:dimnames]{base::dimnames()}} for crochet to work. Implementing methods for
\code{\link[base:nrow]{base::nrow()}}, \code{\link[base:ncol]{base::ncol()}}, \code{\link[base:rownames]{base::rownames()}}, and \code{\link[base:colnames]{base::colnames()}}
is not necessary as the default method of those generics calls \code{\link[base:dim]{base::dim()}}
or \code{\link[base:dimnames]{base::dimnames()}} internally.
}
\examples{
b <- matrix(data = rnorm(25), nrow = 5, ncol = 5)
dimnames(b) <- list(letters[1:5], letters[1:5])

a <- structure(list(), class = "TestMatrix")

dim.TestMatrix <- function(x) c(5, 5)

dimnames.TestMatrix <- function(x) dimnames(b)

subset_vector <- function(x, i) {
        # Dispatch to b instead to x for this demo
        b[i, drop = FALSE]
}

subset_matrix <- function(x, i, j) {
        # Dispatch to b instead to x for this demo
        b[i, j, drop = FALSE]
}

`[.TestMatrix` <- crochet(subset_vector = subset_vector, subset_matrix = subset_matrix)
}
