% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crochet.R
\name{ktoij}
\alias{ktoij}
\title{Convert One-Dimensional Index k to Two-Dimensional Indices i and j}
\usage{
ktoij(x, k)
}
\arguments{
\item{x}{A matrix-like object.}

\item{k}{A one-dimensional index.}
}
\value{
A list containing indices \code{i} and \code{j}.
}
\description{
\code{ktoij} is a helper function that converts a one-dimensional index \code{k} to
two-dimensional indices \code{i} and \code{j}. This can be useful if, for example,
two-dimensional indexing is easier to implement than one-dimensional
indexing.
}
\details{
It is assumed that all indices are one-based.
}
