% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsette.R
\name{subsette}
\alias{subsette}
\title{Creates an implementation of [ for custom matrix-like types}
\usage{
subsette(subset_vector, subset_matrix)
}
\arguments{
\item{subset_vector}{A function in the form of \code{function(x, i)} that
takes a subset of \code{x} based on a single index \code{i} and returns a
vector.}

\item{subset_matrix}{A function in the form of \code{function(x, i, j)} that
takes a subset of \code{x} based on two indices \code{i} and \code{j} and
returns a matrix.}
}
\value{
A function in the form of \code{function(x, i, j, drop = TRUE} that
is meant to be used as a method for \code{[}.
}
\description{
subsette is a function that accepts two arguments \code{subset_vector} (in
the form of \code{function(x, i)}) and \code{subset_matrix} (in the form of
\code{function(x, i, j)}), and returns a function that can be used as a
method for \code{[} for a custom type.
}

