\docType{package}
\name{crochet-package}
\alias{crochet-package}
\title{Implementation Helper for [ and [<- Of Custom Matrix-Like Types}
\description{
crochet is a package that provides functions to help implement the extraction /
subsetting / indexing function \code{[} and replacement function \code{[<-} of
custom matrix-like types (based on S3, S4, etc.), modeled as closely to the
base matrix class as possible (with tests to prove it).
}
\seealso{
\code{\link{extract}} and \code{\link{replace}} to create an implementation of
\code{[} and \code{[<-} for custom types, respecively.
\code{vignette("StringMatrix", package = "crochet")} for a vignette containing
a full example for a custom type.
}
\keyword{package}
